Expanded macros to accept 0 - 5 arguements
There are two problems here. As of this commit :
* When running dbg with multiple args within a function the dbgs are
called in the reverse order. (I.E. Func(dbg(1, 2)) is printed in the
order 2, 1). I believe this is to do with function and might not be
possible to fix.
* Initializer lists no longer work unless surround by brackets at call time.

We will need to update tests.cpp for these things. There is the thought
that we don't need all of these macros and can instead make DebugOutput
into a variadic class. That'll be a new branch if so.